Render projectM in its own OpenGL context.
This is a rather big change to how the plugin works. Rendering for
ProjectM is now done in its own thread, with its own OpenGL
context (that shares resources with the main context, to allow copying
textures back into Rack's window).

This means that information flows between 3 threads: the audio
thread (capturing the PCM data, processing param events like the
button press), the UI thread (taking the texture rendered by the
renderer thread and painting it, adding text, as well as telling the
renderer thread when to change presets) and the renderer
thread (running projectM::renderFrame() in a loop).

This required adding quite a bit of synchronization, which may impact
performance negatively. On my system it doesn't seem to cause issues,
at least.